Refactored garbageCollection.ts code into multiple files (#14220)
Follow up to https://github.com/microsoft/FluidFramework/pull/14209 -
Part 2 of refactoring garbage collection logic.
This PR breaks the logic in garbageCollection.ts into multiple files
keeping only the core logic in garbageCollection.ts.
