+++ packages/runtime/container-runtime/src/summary/index.ts
@@ -76,15 +76,11 @@ export {
 	dataStoreAttributesBlobName,
 	electedSummarizerBlobName,
 	extractSummaryMetadataMessage,
-	GCFeatureMatrix,
-	GCVersion,
-	getGCVersion,
 	getAttributesFormatVersion,
 	getFluidDataStoreAttributes,
 	hasIsolatedChannels,
 	IContainerRuntimeMetadata,
 	ICreateContainerMetadata,
-	IGCMetadata,
 	ISummaryMetadataMessage,
 	metadataBlobName,
 	nonDataStorePaths,
